Add pointers deduced types (#196)
* Add deducing pointer types

Pointers are now deduced after multiple dereferences mixed with
address of operators. If the type is defined in cpp2
it should be deduced. Pointer types are also deduced
when returned from functions.

Deducing of doesn't work for parameter_declaration_list_node.

This change change `sema::get_local_declaration_of` to
`sema::get_declaration_of` that takes additional parameter
that makes it work locally or globally.

* Add tests for deducing pointers

Current limitiation is that when deducing pointers from functions
the order of the functions matters - the deduced functions needs to be
before deduced part of code.

* Add comments to the code